Updated Bento4 to version 1.2.3-387
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1917 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/MP4Splitter/AP4/Source/Core/Ap4Atom.cpp
@@ -297,8 +297,8 @@ AP4_Atom::InspectHeader(AP4_AtomInspector& inspector)
         }
     }
     AP4_FormatString(header, sizeof(header), 
-                     "size=%ld+%lld%s", 
-                     GetHeaderSize(), 
+                     "size=%d+%lld%s", 
+                     (int)GetHeaderSize(), 
                      GetSize()-GetHeaderSize(), 
                      extra);
     inspector.StartElement(name, header);
