+++ src/filters/parser/MP4Splitter/AP4/Source/Core/Ap4DecoderConfigDescriptor.cpp
@@ -132,9 +132,9 @@ AP4_Result
 AP4_DecoderConfigDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
-    AP4_FormatString(info, sizeof(info), "size=%ld+%ld", 
-        GetHeaderSize(),
-        m_PayloadSize);
+    AP4_FormatString(info, sizeof(info), "size=%d+%d", 
+        (int)GetHeaderSize(),
+        (int)m_PayloadSize);
     inspector.StartElement("[DecoderConfig]", info);
     inspector.AddField("stream_type", m_StreamType);
     inspector.AddField("object_type", m_ObjectTypeIndication);
