+++ src/filters/parser/MP4Splitter/AP4/Source/Core/Ap4Expandable.cpp
@@ -116,9 +116,9 @@ AP4_Expandable::Inspect(AP4_AtomInspector& inspector)
     char name[6];
     AP4_FormatString(name, sizeof(name), "[#:%02x]", m_ClassId);
     char info[64];
-    AP4_FormatString(info, sizeof(info), "size=%ld+%ld",
-                     GetHeaderSize(),
-                     m_PayloadSize);
+    AP4_FormatString(info, sizeof(info), "size=%d+%d",
+                     (int)GetHeaderSize(),
+                     (int)m_PayloadSize);
     inspector.StartElement(name, info);
     inspector.EndElement();
