+++ src/filters/parser/MP4Splitter/AP4/Source/Core/Ap4Ipmp.cpp
@@ -85,8 +85,8 @@ AP4_Result
 AP4_IpmpDescriptorPointer::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
-    AP4_FormatString(info, sizeof(info), "size=%ld+%ld", 
-                     GetHeaderSize(),m_PayloadSize);
+    AP4_FormatString(info, sizeof(info), "size=%d+%d", 
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[IPMP_DescriptorPointer]", info);
     inspector.AddField("IPMP_DescriptorID", m_DescriptorId);
     if (m_DescriptorId == 0xFF) {
@@ -201,8 +201,8 @@ AP4_Result
 AP4_IpmpDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
-    AP4_FormatString(info, sizeof(info), "size=%ld+%ld", 
-                     GetHeaderSize(),m_PayloadSize);
+    AP4_FormatString(info, sizeof(info), "size=%d+%d", 
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[IPMP_Descriptor]", info);
     inspector.AddField("IPMP_DescriptorID", m_DescriptorId);
     inspector.AddField("IPMPS_Type", m_IpmpsType, AP4_AtomInspector::HINT_HEX);
